backend/requirements.txt
beautifulsoup4>=4.8.2
lxml>=4.9.0
easyocr>=1.7.2
fastapi>=0.100.0
PyMuPDF>=1.20.0
numpy>=1.22.0
pandas>=2.0.0
Pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.30.0
scipy>=1.10.0
scikit-image>=0.20.0
tiktoken>=0.5.0
torch
torchvision
ultralytics>=8.0.0
uvicorn[standard]>=0.20.0
websockets>=10.4
python-multipart
